backend/requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.4
Django==4.0.3
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
